requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
torch
torchvision
transformers>=4.50.0,<5.0.0
sentence-transformers
timm
Pillow
opencv-python-headless
tqdm
python-dotenv
numpy
qwen-vl-utils
accelerate
peft
